move scripts into split files
